Tweak domain inheritance
